refactor(services): reduce SubBotManager, AIService, PanelServer and shared utils
fix(database): fix DatabaseQueryOptimizer flushBatch key mapping and uncachedKeys calculation
