+++ src/services/download/MediafireDownloader.ts
@@ -8,7 +8,7 @@
 
 import type { Either } from '@/utils/either.js';
 import { left, right } from '@/utils/either.js';
-import { logger } from '@/utils/logger.js';
+import { logger, logError } from '@/utils/logger.js';
 import { NetworkError, ValidationError } from '@/utils/errors.js';
 import fs from 'fs';
 import path from 'path';
@@ -130,7 +130,9 @@ export class MediafireDownloader {
       if (fs.existsSync(filePath)) {
         fs.unlinkSync(filePath);
       }
-    } catch {}
+    } catch (error) {
+      logError('[MediafireDownloader]', error);
+    }
   }
 }
