+++ src/services/permission/LidResolver.ts
@@ -1,6 +1,6 @@
 import type { WASocket } from '@whiskeysockets/baileys';
 import { extractPhone } from './JidService.js';
-import { structuredLogger } from '@/utils/logger.js';
+import { logger } from '@/utils/logger.js';
 
 interface LidPhoneCache {
   get(jid: string): string | undefined;
@@ -21,7 +21,7 @@ export class LidResolver {
         sock as unknown as { onWhatsApp?: (jid: string) => Promise<Array<{ jid?: string }>> }
       ).onWhatsApp;
       if (!onWhatsApp) {
-        structuredLogger.network.warn('[LidResolver] onWhatsApp no disponible', { lidJid });
+        logger.warn({ context: 'LidResolver', message: 'onWhatsApp no disponible', lidJid });
         return null;
       }
 
@@ -32,9 +32,9 @@ export class LidResolver {
         this.lidPhoneCache.set(lidJid, phone);
         return phone;
       }
-      structuredLogger.network.warn('[LidResolver] No se encontró teléfono para LID', { lidJid });
+      logger.warn({ context: 'LidResolver', message: 'No se encontró teléfono para LID', lidJid });
     } catch (error) {
-      structuredLogger.network.warn('[LidResolver] Error resolviendo LID', { lidJid, error });
+      logger.warn({ context: 'LidResolver', message: 'Error resolviendo LID', lidJid, error });
     }
     return null;
   }
