+++ src/services/rpg/data/armor.ts
@@ -1,0 +1,191 @@
+import type { RPGItem } from '../ItemRegistry.js';
+
+export const armor: RPGItem[] = [
+  {
+    id: 'leather_armor',
+    name: 'Armadura de Cuero',
+    description: 'Armadura básica de cuero',
+    type: 'armor',
+    rarity: 'common',
+    stats: { def: 10, agi: 5 },
+    value: 150,
+    sellValue: 75,
+    levelRequired: 1,
+  },
+  {
+    id: 'iron_armor',
+    name: 'Armadura de Hierro',
+    description: 'Armadura de hierro resistente',
+    type: 'armor',
+    rarity: 'common',
+    stats: { def: 25, vit: 5 },
+    value: 400,
+    sellValue: 200,
+    levelRequired: 8,
+  },
+  {
+    id: 'steel_armor',
+    name: 'Armadura de Acero',
+    description: 'Armadura de acero de calidad',
+    type: 'armor',
+    rarity: 'uncommon',
+    stats: { def: 45, vit: 10, hp: 50 },
+    value: 1000,
+    sellValue: 500,
+    levelRequired: 18,
+  },
+  {
+    id: 'golden_armor',
+    name: 'Armadura Dorada',
+    description: 'Armadura legendaria dorada',
+    type: 'armor',
+    rarity: 'epic',
+    stats: { def: 80, vit: 25, hp: 150, luck: 10 },
+    value: 8000,
+    sellValue: 4000,
+    levelRequired: 40,
+  },
+  {
+    id: 'dragon_armor',
+    name: 'Armadura de Dragón',
+    description: 'Armadura forjada con escamas de dragón',
+    type: 'armor',
+    rarity: 'mythic',
+    stats: { def: 150, vit: 50, hp: 300, str: 20 },
+    value: 30000,
+    sellValue: 15000,
+    levelRequired: 70,
+  },
+  {
+    id: 'leather_helmet',
+    name: 'Casco de Cuero',
+    description: 'Casco protector básico',
+    type: 'helmet',
+    rarity: 'common',
+    stats: { def: 5, hp: 20 },
+    value: 80,
+    sellValue: 40,
+    levelRequired: 1,
+  },
+  {
+    id: 'iron_helmet',
+    name: 'Casco de Hierro',
+    description: 'Casco de hierro resistente',
+    type: 'helmet',
+    rarity: 'common',
+    stats: { def: 12, hp: 40 },
+    value: 250,
+    sellValue: 125,
+    levelRequired: 6,
+  },
+  {
+    id: 'steel_helmet',
+    name: 'Casco de Acero',
+    description: 'Casco de acero de calidad',
+    type: 'helmet',
+    rarity: 'uncommon',
+    stats: { def: 20, hp: 80, vit: 5 },
+    value: 600,
+    sellValue: 300,
+    levelRequired: 15,
+  },
+  {
+    id: 'leather_gloves',
+    name: 'Guantes de Cuero',
+    description: 'Guantes protectores básicos',
+    type: 'gloves',
+    rarity: 'common',
+    stats: { def: 3, atk: 3 },
+    value: 50,
+    sellValue: 25,
+    levelRequired: 1,
+  },
+  {
+    id: 'iron_gloves',
+    name: 'Guantes de Hierro',
+    description: 'Guantes de hierro para combate',
+    type: 'gloves',
+    rarity: 'common',
+    stats: { def: 8, atk: 8 },
+    value: 180,
+    sellValue: 90,
+    levelRequired: 5,
+  },
+  {
+    id: 'leather_boots',
+    name: 'Botas de Cuero',
+    description: 'Botas cómodas para caminar',
+    type: 'boots',
+    rarity: 'common',
+    stats: { def: 4, agi: 5 },
+    value: 60,
+    sellValue: 30,
+    levelRequired: 1,
+  },
+  {
+    id: 'iron_boots',
+    name: 'Botas de Hierro',
+    description: 'Botas protectoras resistentes',
+    type: 'boots',
+    rarity: 'common',
+    stats: { def: 10, agi: 8 },
+    value: 200,
+    sellValue: 100,
+    levelRequired: 7,
+  },
+  {
+    id: 'speed_boots',
+    name: 'Botas de Velocidad',
+    description: 'Botas que aumentan la velocidad',
+    type: 'boots',
+    rarity: 'rare',
+    stats: { def: 15, agi: 20, dodgeChance: 10 },
+    value: 1500,
+    sellValue: 750,
+    levelRequired: 20,
+  },
+  {
+    id: 'lucky_charm',
+    name: 'Amuleto de la Suerte',
+    description: 'Un amuleto que trae buena suerte',
+    type: 'accessory',
+    rarity: 'uncommon',
+    stats: { luck: 15 },
+    value: 500,
+    sellValue: 250,
+    levelRequired: 10,
+  },
+  {
+    id: 'power_ring',
+    name: 'Anillo de Poder',
+    description: 'Un anillo que aumenta el poder',
+    type: 'accessory',
+    rarity: 'rare',
+    stats: { atk: 25, str: 10 },
+    value: 1200,
+    sellValue: 600,
+    levelRequired: 15,
+  },
+  {
+    id: 'protection_ring',
+    name: 'Anillo de Protección',
+    description: 'Un anillo que protege',
+    type: 'accessory',
+    rarity: 'rare',
+    stats: { def: 20, vit: 15 },
+    value: 1200,
+    sellValue: 600,
+    levelRequired: 15,
+  },
+  {
+    id: 'wisdom_amulet',
+    name: 'Amuleto de Sabiduría',
+    description: 'Un amuleto de conocimiento',
+    type: 'accessory',
+    rarity: 'rare',
+    stats: { int: 25, critChance: 10 },
+    value: 1800,
+    sellValue: 900,
+    levelRequired: 20,
+  },
+];
