+++ src/services/rpg/data/consumables.ts
@@ -1,0 +1,148 @@
+import type { RPGItem } from '../ItemRegistry.js';
+
+export const consumables: RPGItem[] = [
+  {
+    id: 'health_potion_small',
+    name: 'Poción Pequeña de Vida',
+    description: 'Restaura 30 HP',
+    type: 'consumable',
+    rarity: 'common',
+    stats: {},
+    value: 25,
+    sellValue: 10,
+    levelRequired: 1,
+    useEffect: { type: 'heal', value: 30 },
+  },
+  {
+    id: 'health_potion_medium',
+    name: 'Poción Mediana de Vida',
+    description: 'Restaura 75 HP',
+    type: 'consumable',
+    rarity: 'common',
+    stats: {},
+    value: 75,
+    sellValue: 35,
+    levelRequired: 5,
+    useEffect: { type: 'heal', value: 75 },
+  },
+  {
+    id: 'health_potion_large',
+    name: 'Poción Grande de Vida',
+    description: 'Restaura 150 HP',
+    type: 'consumable',
+    rarity: 'uncommon',
+    stats: {},
+    value: 200,
+    sellValue: 100,
+    levelRequired: 15,
+    useEffect: { type: 'heal', value: 150 },
+  },
+  {
+    id: 'health_potion_super',
+    name: 'Poción Super de Vida',
+    description: 'Restaura 300 HP',
+    type: 'consumable',
+    rarity: 'rare',
+    stats: {},
+    value: 500,
+    sellValue: 250,
+    levelRequired: 30,
+    useEffect: { type: 'heal', value: 300 },
+  },
+  {
+    id: 'energy_potion',
+    name: 'Poción de Energía',
+    description: 'Restaura 50 de energía',
+    type: 'consumable',
+    rarity: 'common',
+    stats: {},
+    value: 40,
+    sellValue: 20,
+    levelRequired: 1,
+    useEffect: { type: 'restoreEnergy', value: 50 },
+  },
+  {
+    id: 'stamina_potion',
+    name: 'Poción de Resistencia',
+    description: 'Restaura 50 de resistencia',
+    type: 'consumable',
+    rarity: 'common',
+    stats: {},
+    value: 35,
+    sellValue: 15,
+    levelRequired: 1,
+    useEffect: { type: 'restoreStamina', value: 50 },
+  },
+  {
+    id: 'bread',
+    name: 'Pan',
+    description: 'Restaura 10 HP',
+    type: 'consumable',
+    rarity: 'common',
+    stats: {},
+    value: 10,
+    sellValue: 5,
+    levelRequired: 1,
+    useEffect: { type: 'heal', value: 10 },
+  },
+  {
+    id: 'meat',
+    name: 'Carne',
+    description: 'Restaura 25 HP',
+    type: 'consumable',
+    rarity: 'common',
+    stats: {},
+    value: 20,
+    sellValue: 10,
+    levelRequired: 1,
+    useEffect: { type: 'heal', value: 25 },
+  },
+  {
+    id: 'fish',
+    name: 'Pescado',
+    description: 'Restaura 20 HP',
+    type: 'consumable',
+    rarity: 'common',
+    stats: {},
+    value: 15,
+    sellValue: 7,
+    levelRequired: 1,
+    useEffect: { type: 'heal', value: 20 },
+  },
+  {
+    id: 'cookie',
+    name: 'Galleta',
+    description: 'Restaura 15 HP',
+    type: 'consumable',
+    rarity: 'common',
+    stats: {},
+    value: 12,
+    sellValue: 5,
+    levelRequired: 1,
+    useEffect: { type: 'heal', value: 15 },
+  },
+  {
+    id: 'apple',
+    name: 'Manzana',
+    description: 'Restaura 12 HP',
+    type: 'consumable',
+    rarity: 'common',
+    stats: {},
+    value: 8,
+    sellValue: 3,
+    levelRequired: 1,
+    useEffect: { type: 'heal', value: 12 },
+  },
+  {
+    id: 'xp_scroll',
+    name: 'Pergamino de XP',
+    description: 'Da 100 XP extra',
+    type: 'consumable',
+    rarity: 'uncommon',
+    stats: {},
+    value: 150,
+    sellValue: 75,
+    levelRequired: 5,
+    useEffect: { type: 'xpBoost', value: 100 },
+  },
+];
