+++ src/services/rpg/data/materials.ts
@@ -1,0 +1,136 @@
+import type { RPGItem } from '../ItemRegistry.js';
+
+export const materials: RPGItem[] = [
+  {
+    id: 'wood',
+    name: 'Madera',
+    description: 'Madera básica para crafting',
+    type: 'material',
+    rarity: 'common',
+    stats: {},
+    value: 5,
+    sellValue: 2,
+    levelRequired: 1,
+  },
+  {
+    id: 'stone',
+    name: 'Piedra',
+    description: 'Piedra para crafting',
+    type: 'material',
+    rarity: 'common',
+    stats: {},
+    value: 8,
+    sellValue: 3,
+    levelRequired: 1,
+  },
+  {
+    id: 'iron_ore',
+    name: 'Mineral de Hierro',
+    description: 'Mineral de hierro refinado',
+    type: 'material',
+    rarity: 'common',
+    stats: {},
+    value: 15,
+    sellValue: 7,
+    levelRequired: 1,
+  },
+  {
+    id: 'silver_ore',
+    name: 'Mineral de Plata',
+    description: 'Mineral de plata valiosa',
+    type: 'material',
+    rarity: 'uncommon',
+    stats: {},
+    value: 40,
+    sellValue: 20,
+    levelRequired: 10,
+  },
+  {
+    id: 'gold_ore',
+    name: 'Mineral de Oro',
+    description: 'Mineral de oro precioso',
+    type: 'material',
+    rarity: 'rare',
+    stats: {},
+    value: 100,
+    sellValue: 50,
+    levelRequired: 20,
+  },
+  {
+    id: 'diamond',
+    name: 'Diamante',
+    description: 'Una gema brillante',
+    type: 'material',
+    rarity: 'epic',
+    stats: {},
+    value: 500,
+    sellValue: 250,
+    levelRequired: 30,
+  },
+  {
+    id: 'ruby',
+    name: 'Rubí',
+    description: 'Una gema roja',
+    type: 'material',
+    rarity: 'rare',
+    stats: {},
+    value: 300,
+    sellValue: 150,
+    levelRequired: 25,
+  },
+  {
+    id: 'emerald',
+    name: 'Esmeralda',
+    description: 'Una gema verde',
+    type: 'material',
+    rarity: 'rare',
+    stats: {},
+    value: 300,
+    sellValue: 150,
+    levelRequired: 25,
+  },
+  {
+    id: 'leather',
+    name: 'Cuero',
+    description: 'Cuero de animal',
+    type: 'material',
+    rarity: 'common',
+    stats: {},
+    value: 20,
+    sellValue: 10,
+    levelRequired: 1,
+  },
+  {
+    id: 'magic_dust',
+    name: 'Polvo Mágico',
+    description: 'Polvo con propiedades mágicas',
+    type: 'material',
+    rarity: 'uncommon',
+    stats: {},
+    value: 50,
+    sellValue: 25,
+    levelRequired: 10,
+  },
+  {
+    id: 'dragon_scale',
+    name: 'Escama de Dragón',
+    description: 'Una escama de dragón',
+    type: 'material',
+    rarity: 'legendary',
+    stats: {},
+    value: 1000,
+    sellValue: 500,
+    levelRequired: 50,
+  },
+  {
+    id: 'phoenix_feather',
+    name: 'Pluma de Fénix',
+    description: 'Una pluma de fénix',
+    type: 'material',
+    rarity: 'mythic',
+    stats: {},
+    value: 2000,
+    sellValue: 1000,
+    levelRequired: 60,
+  },
+];
