+++ src/services/rpg/data/weapons.ts
@@ -1,0 +1,169 @@
+import type { RPGItem } from '../ItemRegistry.js';
+
+export const weapons: RPGItem[] = [
+  {
+    id: 'wooden_sword',
+    name: 'Espada de Madera',
+    description: 'Una espada básica feita de madera',
+    type: 'weapon',
+    rarity: 'common',
+    stats: { atk: 5 },
+    value: 50,
+    sellValue: 25,
+    levelRequired: 1,
+  },
+  {
+    id: 'iron_sword',
+    name: 'Espada de Hierro',
+    description: 'Una espada de hierro resistente',
+    type: 'weapon',
+    rarity: 'common',
+    stats: { atk: 15 },
+    value: 200,
+    sellValue: 100,
+    levelRequired: 5,
+  },
+  {
+    id: 'steel_sword',
+    name: 'Espada de Acero',
+    description: 'Una espada de acero de buena calidad',
+    type: 'weapon',
+    rarity: 'uncommon',
+    stats: { atk: 30, critChance: 5 },
+    value: 500,
+    sellValue: 250,
+    levelRequired: 10,
+  },
+  {
+    id: 'silver_sword',
+    name: 'Espada de Plata',
+    description: 'Una espada encantada con plata',
+    type: 'weapon',
+    rarity: 'rare',
+    stats: { atk: 50, luck: 10 },
+    value: 1500,
+    sellValue: 750,
+    levelRequired: 20,
+  },
+  {
+    id: 'golden_sword',
+    name: 'Espada Dorada',
+    description: 'Una legendaria espada dorada',
+    type: 'weapon',
+    rarity: 'epic',
+    stats: { atk: 80, critChance: 15, luck: 20 },
+    value: 5000,
+    sellValue: 2500,
+    levelRequired: 35,
+  },
+  {
+    id: 'diamond_sword',
+    name: 'Espada de Diamante',
+    description: 'La espada más dura conocida',
+    type: 'weapon',
+    rarity: 'legendary',
+    stats: { atk: 120, critChance: 20, luck: 30 },
+    value: 15000,
+    sellValue: 7500,
+    levelRequired: 50,
+  },
+  {
+    id: 'dragon_sword',
+    name: 'Espada de Dragón',
+    description: 'Forjada con escamas de dragón',
+    type: 'weapon',
+    rarity: 'mythic',
+    stats: { atk: 200, critChance: 30, luck: 50, str: 20 },
+    value: 50000,
+    sellValue: 25000,
+    levelRequired: 75,
+  },
+  {
+    id: 'wooden_bow',
+    name: 'Arco de Madera',
+    description: 'Un arco básico de entrenamiento',
+    type: 'weapon',
+    rarity: 'common',
+    stats: { atk: 8, agi: 5 },
+    value: 75,
+    sellValue: 35,
+    levelRequired: 1,
+  },
+  {
+    id: 'hunters_bow',
+    name: 'Arco del Cazador',
+    description: 'Un arco usado por cazadores expertos',
+    type: 'weapon',
+    rarity: 'uncommon',
+    stats: { atk: 25, agi: 15 },
+    value: 600,
+    sellValue: 300,
+    levelRequired: 12,
+  },
+  {
+    id: 'elven_bow',
+    name: 'Arco Élfico',
+    description: 'Un arco élfico de gran precisión',
+    type: 'weapon',
+    rarity: 'rare',
+    stats: { atk: 55, agi: 25, critChance: 10 },
+    value: 2000,
+    sellValue: 1000,
+    levelRequired: 25,
+  },
+  {
+    id: 'wooden_staff',
+    name: 'Bastón de Madera',
+    description: 'Un bastón para mago novato',
+    type: 'weapon',
+    rarity: 'common',
+    stats: { atk: 6, int: 8 },
+    value: 60,
+    sellValue: 30,
+    levelRequired: 1,
+  },
+  {
+    id: 'arcane_staff',
+    name: 'Bastón Arcano',
+    description: 'Un bastón que canaliza poder mágico',
+    type: 'weapon',
+    rarity: 'uncommon',
+    stats: { atk: 20, int: 20, mana: 50 },
+    value: 700,
+    sellValue: 350,
+    levelRequired: 15,
+  },
+  {
+    id: 'void_staff',
+    name: 'Bastón del Vacío',
+    description: 'Un bastón de poder oscuro',
+    type: 'weapon',
+    rarity: 'epic',
+    stats: { atk: 70, int: 45, critChance: 15 },
+    value: 6000,
+    sellValue: 3000,
+    levelRequired: 40,
+  },
+  {
+    id: 'dagger',
+    name: 'Daga',
+    description: 'Una daga rápida y afilada',
+    type: 'weapon',
+    rarity: 'common',
+    stats: { atk: 10, agi: 10 },
+    value: 100,
+    sellValue: 50,
+    levelRequired: 3,
+  },
+  {
+    id: 'shadow_dagger',
+    name: 'Daga de las Sombras',
+    description: 'Una daga que Grants invisibilidad',
+    type: 'weapon',
+    rarity: 'rare',
+    stats: { atk: 45, agi: 30, dodgeChance: 15 },
+    value: 2500,
+    sellValue: 1250,
+    levelRequired: 22,
+  },
+];
