+++ src/services/system/ResilienceService.ts
@@ -63,18 +63,6 @@ function clampNumber(value: number, min: number, max: number, fallback: number):
   return Math.min(max, Math.max(min, n));
 }
 
-function formatDuration(ms: number): string {
-  const seconds = Math.floor(ms / 1000);
-  const minutes = Math.floor(seconds / 60);
-  const hours = Math.floor(minutes / 60);
-  const days = Math.floor(hours / 24);
-
-  if (days > 0) return `${days}d ${hours % 24}h`;
-  if (hours > 0) return `${hours}h ${minutes % 60}m`;
-  if (minutes > 0) return `${minutes}m`;
-  return `${seconds}s`;
-}
-
 export class ResilienceService {
   private store: ResilienceStore;
 
@@ -200,4 +188,4 @@ export class ResilienceService {
 }
 
 export const resilienceService = new ResilienceService();
-export { formatDuration };
+export { formatTime as formatDuration } from '@/utils/helpers.js';
