+++ src/utils/constants.ts
@@ -47,11 +47,18 @@ export const DATABASE = {
 
 export const AUTH = {
   MAX_QR_RETRIES: 10,
-  MAX_RECONNECT_ATTEMPTS: 15,
-  MAX_RECONNECT_DELAY: 5 * TIME.SECOND,
+  MAX_RECONNECT_ATTEMPTS: 20,
+  MAX_RECONNECT_DELAY: 5 * 60_000,
+  FIRST_RECONNECT_DELAY: 15_000,
+  RECONNECT_BASE_DELAY: 1_000,
   ERROR_515_MAX_RETRIES: 3,
   PAIRING_CODE_EXPIRY: 3 * TIME.MINUTE,
 } as const;
+
+export const MAX_RECONNECT_ATTEMPTS = 20;
+export const MAX_RECONNECT_DELAY = 5 * 60_000;
+export const FIRST_RECONNECT_DELAY = 15_000;
+export const RECONNECT_BASE_DELAY = 1_000;
 
 export const MAINTENANCE = {
   CLEANUP_INTERVAL: TIME.HOUR,
