+++ src/utils/errors.ts
@@ -32,6 +32,11 @@ export enum ErrorCode {
   GROUP_NOT_FOUND = 'GROUP_NOT_FOUND',
   INTERNAL_ERROR = 'INTERNAL_ERROR',
   SERVICE_UNAVAILABLE = 'SERVICE_UNAVAILABLE',
+  AI_ERROR = 'AI_ERROR',
+  DOWNLOAD_ERROR = 'DOWNLOAD_ERROR',
+  MEDIA_ERROR = 'MEDIA_ERROR',
+  DATABASE_ERROR = 'DATABASE_ERROR',
+  USER_BANNED = 'USER_BANNED',
 }
 
 export class VBotError extends Error {
